Trait ceres_launchpad::pallet::Config

pub trait Config: Config + Config + Config + Config + Config + Config + Config + Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type WeightInfo: WeightInfo;

    const MILLISECONDS_PER_DAY: Self::Moment;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Required Associated Constants

const MILLISECONDS_PER_DAY: Self::Moment

One day represented in milliseconds